+++ CHANGELOG.md
@@ -10,6 +10,10 @@ ENHANCEMENTS:
 * resource/`junos_ospf_area`: add `realm` argument to add a OSPFv3 realm area. (Fixes #392),  
   add new `area_range`, `context_identifier`, `inter_area_prefix_export`, `inter_area_prefix_import`, `network_summary_export`, `network_summary_import`, `no_context_identifier_advertisement`, `nssa`, `stub` and `virtual_link` arguments,  
   add validation of `area_id`
+* resource/`junos_interface_logical`: add `disable` argument (Fixes #395)
+* data-source/`junos_interface_logical`: add `disable` attribute (like resource)
+* resource/`junos_interface_physical`: add `disable` argument
+* data-source/`junos_interface_physical`: add `disable` attribute (like resource)
 
 BUG FIXES: